chore(org. administration): restructured and combined invite user and member list
